Build system: Add a formulaic setting of Q from V for debugging
Build system: Add a formulaic setting of Q from V for debugging

We want to be able to disable the many @'s in the Makefiles so
that we can see what they are doing, when things go wrong.  It is
conventional to arrange for setting V=1 to have this effect.

Here we supply a formulaic stanza for either setting Q?=@ or
doing nothing.  (There is sadly no standard place which is
included in all the Makefiles so this is probably best).

In this patch we do not introduce any users of Q yet.  This is
because the next patch, which introduces all the users of Q, can
be generated entirely automatically.  (This is also convenient in
case something needs to be rebased across it.)

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>