Build system: Fix ./build debian
Build system: Fix ./build debian

We need to -I and -L the build tree directories, not the directories
in /usr/local or wheveever.  Otherwise (a) the `./debian build' fails
if wiringPi is not installed (b) if it _is_ installed, the build picks
up the installed versions rather than the versions being built, which
is wrong.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>