build: Introduce a variable "$sudo" and honour WIRINGPI_SUDO
build: Introduce a variable "$sudo" and honour WIRINGPI_SUDO

We are going to want to not run sudo sometimes.  So replace literal
invocations of sudo with a variable, which is set to `sudo' by
default, but can be overridden from the environment.

No functional change if the environment variable is not set.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>