Build system: Move template-based debian build into debian-template
Build system: Move template-based debian build into debian-template

We are going to want to introduce a real Debian source package build
system.  But that needs to contain a lot of things in debian/ which
ought not to appear in the .deb made by the existing ad-hoc `./build
debian' arrangements, which we want to keep so as not to unduly
disturb existing users.

So rename the whole debian/ directory to debian-template/ and change
all references in the build system.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>